VersionGuesser: Check all git tags for current HEAD
..in case `git describe --exact-match --tags`
doesn't return a normalizable version.
Which may be the case when one commit has several
tags.

This change uses `git tag --point-at`
which is available as of git 1.7.0

TODO: Check if this runs smooth
(it should) with earlier version
TODO: Check if explode with "\n" or PHP_EOL should
be used for win compatibility.

Example:
$ git describe --exact-match --tags
TYPO3_8-7-5
$ git tag --list --points-at HEAD
8.7.5
TYPO3_8-7-5
v8.7.5